+++ android/src/main/java/io/parity/signer/screens/scan/ScanNavSubgraph.kt
@@ -81,7 +81,6 @@ fun ScanNavSubgraph(
 				scanViewModel.errorWrongPassword.value = true
 				scanViewModel.bananaSplitPassword.value = null
 			},
-			modifier = Modifier.statusBarsPadding(),
 		)
 	} else if (transactionsValue == null || showingModals) {
